FastAPI-Kafka-Topic-Creation/requirements.txt
kafka-python==2.0.2
fastapi[all]==0.75.0
python-dotenv==0.20.0